requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
filelock==3.14.0
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.7
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pillow==10.3.0
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.3
sqlparse==0.5.0
sympy==1.12
timm==0.9.16
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
